# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Computer‑Based Testing API with subscriptions, academic management, and CBT modules.",
        "title": "CBT API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@cbt-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/class-levels": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a class level (e.g., \"Grade 10\") for a school.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Academic"
                ],
                "summary": "Create a new class level",
                "parameters": [
                    {
                        "description": "Class level details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateClassLevelRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "message + data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/class-levels/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a single class level.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Academic"
                ],
                "summary": "Get class level by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Class Level ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Modify an existing class level.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Academic"
                ],
                "summary": "Update a class level",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Class Level ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateClassLevelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message + data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Soft‑delete a class level.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Academic"
                ],
                "summary": "Delete a class level",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Class Level ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/school-class-levels/{schoolId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List all class levels belonging to a school.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Academic"
                ],
                "summary": "Get all class levels for a school",
                "parameters": [
                    {
                        "type": "string",
                        "description": "School ID",
                        "name": "schoolId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data (list)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateClassLevelRequest": {
            "type": "object",
            "required": [
                "category",
                "level_number",
                "name",
                "school_id"
            ],
            "properties": {
                "category": {
                    "type": "string",
                    "enum": [
                        "JSS",
                        "SSS",
                        "PRIMARY"
                    ]
                },
                "level_number": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "promotion_to": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateClassLevelRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "level_number": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "promotion_to": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter **Bearer \u0026lt;token\u0026gt;**. Example: `Bearer eyJhbGciOiJIUzI1NiIs...`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}